Remove irrelevant changes
